converting to modular
not sure how I feel about this yet.

Right now the code is jacked up because instead of using globals for
common things I have them stored as class variables and haven't passed
them around to all the functions yet.

The code was very procedurally written so making use of a module this
late seems like it's just making a mess.  But I can always back out if I
change my mind.  This should allow us to unit test parts of the code and
I hope the result will be cleaner than what I started with.  If it isn't
that may be another reason I drop it. :)