everything is working again
I still need to do lots of tests and start writing some of the tests as
unit tests.  Specifically I'm not sure if handlers work quite right
again, and things like aliases.

The code is also still really complex where it shouldn't be, but I think
it may be getting better.