IO::Stty is required
Expect really needed it but had it as optional.  Interactive terminals
really don't work right without being able to turn on and off stty.

We also use it directly in fake-a-router.  I thought Expect required it
so I never specified it as a dependency, but that was causing the tests
to fail almost everywhere.