STDIN not a tty now dies with error
things like cat routers | xargs tel will die with an understandable
error if a tty is required (maybe.. there maybe other places where input
isn't checked, like if it makes it to the interact stage)

On freebsd you can fix this with their xargs by reopening stdin.  You
might be able to fix it in linux with a similar hack, but it's not
something the script needs to worry about.. we just need to exit if the
tty can't handle input.