changed the passwd module to be more strict
I started out trying to incorporate Module::Pluggable but my
architecture is kind of weird for that.  I did notice that mapping was
kind of pointless though so I removed it and replaced it with a line
that converts to lc($plugin).  This breaks some assumptions that really
only should have been used in tests, so I think it's fine.