very much improved the error reporting for passwd
most of the modules croak on errors rather than silently discarding
them.  My original thought was that since modules were optional we
should discard errors, but masking them was creating a headache for me
and probably for users as well.

The truth is that once you get to a point of seeing those errors you
would need them (because load_module shouldn't be called until you are
using one of the optional modules)