big changes to help unit testing
This completely rewrites the fake-a-router stuff to be more gooder.  We
fixed a couple of obscure bugs along the way too.. then we put some
shitty code into control_loop because we aren't handling eval/die with
try/catch..

but we can't have everything.

control_loop probably needs a complete rewrite to fix the error
handling.