fix connection handler problem
This only affects logins with ssh keys, and strangely I only could get
it to happen if there was no controlling tty.

The issue is that there ends up being two expects for the prompt.  The
first one determines that you have logged in, the second is before
sending the pagercmd.

This fix added a dependency (enum) but I think it's cleaner than the
alternative.