fix bad password handling when nologin set.
Normally if you have the wrong password it will prompt again but
consider you to be connected.  This is a feature for me because it
allows me to fix things in interactive sessions.

The downside is that it causes a permanent hang and ignores the timeout
when in a non-interactive session.

This is a workaround for no-login sessions.  I might revisit this later
and try to detect password failures vs a valid password prompt.  I'm not
sure how it could be better handled right now.