workaround for mikrotiks with ssh key login
This is a workaround for the ANSI detection on mikrotiks.  If you don't
have something like this then it can hang for up to 10 seconds and then
will fail to use ANSI colors.

If a password prompt happens then this isn't an issue.  The password is
sent, connected(CONN_PASSWORD) is set and we exit the login method.

When an ssh key is provided this skips the password prompt so we're
still in the main login loop.  In theory you could wait for the prompt,
which is what we're doing, but the prompt doesn't appear until after the
ANSI detection.

Another solution is to detect the ANSI detection and provide answers,
but this would probably screw up the state between the tty and the
remote side.  The ideal here would be to detect the ANSI detection, pass
it on to ssh/the tty app to let them respond.  We could do this by
dropping to connected mode, but we have no way to be sure we're talking
to a mikrotik, or where we are in the login sequence.  So if someone sent
an ANSI detect code before a password prompt it would kick the session out
to connected mode.

So the workaround is to watch for a banner line and if we see it we know
we're already logged in and can exit the login sequence.