add %%perleval to run_commands
This needs to be documented and probably flagged so that if it's not
enabled in a config file it's not possible to do.

The idea is if you have a script to execute but you need it to wait (or
do something else) then you can insert a sleep (or other command) by
typing:

%%perleval: sleep 20

Inside of the script, and it will not be sent to the remote device.
Instead it will be run by the local perl script.