apply major changes described below
I haven't worked on this script much in the last 10 years or so.  That
isn't because it's not used, it's more because it did everything I
needed it to do and worked well enough to not bother fixing the
problems.

I don't use Perl much anymore.  That is not because of the language,
which I still think is nice.  It's more because the ecosystem isn't
there as much.  If I need to interface with some external thing,
frequently there is a python module already written, where I would need
to write my own module for perl.  Because of this I found myself working
on tel even less.

These are changes that are leading up to a 1.0.0 release.  Most of them
involve removing features around authentication and color.

I'm modifying the github workflow to automatically build a docker
version.  This is an experiment.  It might not work well because of
permissions for the telrc files mismatching with the docker container.
This may not work out.  I'll need to see if there is a compatible
workflow.  It would be nice if we could just install an alias for tel
that was "docker run ghcr.io/rfdrake/tel"

The code for colorize has been removed.  The idea being
that an external colorizer like App-RouterColorizer does a better job.
The only drawback to RouterColorizer (besides the busy loop it needs to
function correctly) is that it would be nice to have a modular config
that is easy to customize.

The perl modules I had within tel were not that.  Hopefully I can work
on something that will do the colorizer stuff and supports a modular
config so people can easily write extensions for it.

The authentication code is being removed and documentation being
updated.  This removes support for things like Keypass/Keyring and other
password managers.  The reason for this is twofold:

1.  Supporting modules that are not tested very much and not used by me
seems like a bad idea.  There may be broken stuff or security problems
that I don't notice for years because I never use that feature.

2.  Most of what I was trying to do can be accomplished in a couple of
ways.  Environment variable forwarding for ssh along with router tacacs
passwords in an environment variable takes care of most of the security
issues.  At that point you make sure your laptop has an encrypted
partition and the security is handled by not storing the passwords on a
shared jumphost.  There are edge cases, like individual passwords for
each router stored in a safe.  Those can be handled by forking an
executable that returns the password.  So, if say bitwarden was your
password manager, then you would invoke "bwcli" to pull the appropriate
password.  This will be documented in an AUTHENTICATION.md file.

There are several other changes that are happening.  I'm just
documenting this commit about 2 years after staging the changes, so I
probably forgot some things.