remove _stty_rows, add connection_method
This removes _stty_rows.  It was an experiment that I never got to work. The idea was that we could tell a device that we were a 3000 line long terminal and defeat the devices "More" prompt, but no device that I found would accept such a thing. Most devices have a hard coded screen size of xxx lines which shows a more prompt even if it thinks the physical screen is bigger.

For the "connection_method" I decided I wanted a better way to determine the hostname or IP of the device I connect to when parsing script output. Usually I look for the line "Adding this thing to known hosts.." or whatever, but that doesn't work if the device is already in known hosts, or you use telnet, or any number of other things.  So now if you turn on --debug it will show a connection string.